Update rockPaperScissors.py
Changed game logic to be simpler, added a balance system to bet on the outcome. Balance will be transferable to future games i create. I generally redid a lot of the game to work better. I'm still running into errors with the balance file sometimes bricking so I might have to update the game later but it's good for now.